+++ activesupport/lib/active_support/testing/autorun.rb
@@ -7,5 +7,11 @@ require "minitest"
 # unbundled environment, which triggers MT5 usage. This conditional may
 # be removable after the version bump, though it currently safeguards
 # against issues in environments with multiple versions installed.
-Minitest.load :rails if Minitest.respond_to? :load
+if Minitest.respond_to? :load
+  # Auto-load all installed plugins when using Minitest 6, letting the "rails"
+  # plugin override CLI flags last.
+  Minitest.load_plugins
+  Minitest.extensions.delete("rails")
+  Minitest.load :rails
+end
 Minitest.autorun